+++ mltu/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.1.1"
+__version__ = "1.1.2"
 
 from .annotations.images import Image
 from .annotations.images import CVImage